fix/fix-assets-folder (#151)
* fix: fix assets used in all screens

* fix: fix assets used in all screens

* refactor: extract close button to a specific widget
