+++ lib/overlays/admin_overlay.dart
@@ -6,6 +6,7 @@ import 'package:dextraquario/fish_info.dart';
 import 'package:flame/widgets/nine_tile_box.dart';
 import 'package:flame/widgets/sprite_button.dart';
 import 'package:flutter/material.dart';
+import 'package:dextraquario/components/close_button_widget.dart';
 
 import '../contribution.dart';
 
@@ -28,16 +29,7 @@ class AdminOverlay extends StatelessWidget {
           Row(
             mainAxisAlignment: MainAxisAlignment.end,
             children: [
-              Padding(
-                padding: EdgeInsets.only(top: 44, right: 44),
-                child: SpriteButton(
-                    onPressed: () => onClose?.call(),
-                    label: null,
-                    width: 48,
-                    height: 48,
-                    sprite: Assets.closeButton48,
-                    pressedSprite: Assets.closeButton48),
-              ),
+              CloseButtonWidget(onClick: onClose),
             ],
           ),
           Column(
