+++ lib/overlays/profile_overlay.dart
@@ -1,3 +1,4 @@
+import 'package:dextraquario/components/close_button_widget.dart';
 import 'package:dextraquario/fish_info.dart';
 import 'package:dextraquario/models/user_model.dart';
 import 'package:dextraquario/services/user_service.dart';
@@ -75,18 +76,12 @@ class ProfileScreen extends StatelessWidget {
 
     return Stack(
       children: [
-        Row(mainAxisAlignment: MainAxisAlignment.end, children: [
-          Padding(
-            padding: EdgeInsets.only(top: 44, right: 44),
-            child: SpriteButton(
-                onPressed: () => onClose?.call(),
-                label: null,
-                width: 48,
-                height: 48,
-                sprite: Assets.closeButton48,
-                pressedSprite: Assets.closeButton48),
-          ),
-        ]),
+        Row(
+          mainAxisAlignment: MainAxisAlignment.end,
+          children: [
+            CloseButtonWidget(onClick: onClose),
+          ],
+        ),
         Column(
           mainAxisAlignment: MainAxisAlignment.center,
           crossAxisAlignment: CrossAxisAlignment.center,
